requirements.txt
alembic==1.12.1
asyncpg==0.29.0
fastapi[all]==0.104.1
fastapi-users[sqlalchemy]==12.1.2
fastapi-mail==1.4.1
psycopg2-binary==2.9.9
bcrypt==4.0.1
gunicorn==21.2.0
aiofiles==23.2.1
fastapi-pagination==0.12.12
# httpx-oauth
fastapi-cache2==0.2.0
redis==4.4.2
sqladmin[full]==0.16.0
pytest-asyncio==0.23.4
sentry-sdk[fastapi]==1.39.2
